requirements.txt
pubsubpy
celery

